Move amis integration into main project
This allows more easy running of the tests from the root directory and is more consistent project structure
